refactor: Move BoundingBoxClipper and use module import paths
Please enter the commit message for your changes. Lines starting
