+++ viewer/core/src/index.ts
@@ -11,9 +11,9 @@
  */
 
 export { revealEnv, IndexSet, NumericRange } from '@reveal/utilities';
+export { BoundingBoxClipper } from '@reveal/cad-model';
 
 export * from './public/migration/Cognite3DViewer';
-export { BoundingBoxClipper } from './utilities';
 export { Cognite3DModel } from './public/migration/Cognite3DModel';
 export { Cognite3DViewer } from './public/migration/Cognite3DViewer';
 export { CognitePointCloudModel } from './public/migration/CognitePointCloudModel';
