+++ viewer/core/src/public/types.ts
@@ -2,24 +2,7 @@
  * Copyright 2021 Cognite AS
  */
 
-import { LoadingState, SectorCuller } from '@reveal/cad-geometry-loaders';
-import { SectorQuads, RenderOptions } from '@reveal/rendering';
-import { SectorGeometry } from '@reveal/cad-parsers';
-
-/**
- * @property logMetrics Might be used to disable usage statistics.
- * @property nodeAppearanceProvider Style node by tree-index.
- * @property internal Internals are for internal usage only (like unit-testing).
- */
-export type RevealOptions = {
-  logMetrics?: boolean;
-  renderOptions?: RenderOptions;
-  continuousModelStreaming?: boolean;
-  internal?: {
-    parseCallback?: (parsed: { lod: string; data: SectorGeometry | SectorQuads }) => void;
-    sectorCuller?: SectorCuller;
-  };
-};
+import { LoadingState } from '@reveal/cad-geometry-loaders';
 
 /**
  * Handler for events about data being loaded.
