+++ viewer/packages/cad-model/index.ts
@@ -5,3 +5,6 @@
 export { CadManager } from './src/CadManager';
 export { GeometryFilter } from './src/GeometryFilter';
 export { CadModelSectorLoadStatistics } from './src/CadModelSectorLoadStatistics';
+export { intersectCadNodes } from './src/picking';
+export { createCadManager } from './src/createCadManager';
+export { BoundingBoxClipper } from './src/BoundingBoxClipper';
