+++ viewer/packages/cad-model/src/BoundingBoxClipper.test.ts
