+++ viewer/packages/cad-model/src/BoundingBoxClipper.ts
