+++ viewer/packages/model-base/index.ts
@@ -2,5 +2,5 @@
  * Copyright 2022 Cognite AS
  */
 
-export { IntersectInput } from './src/types';
+export { RevealOptions, IntersectInput } from './src/types';
 export { SupportedModelTypes } from './src/SupportedModelTypes';
